Correctly output program versions
Have this Action correctly set the program versions as Action outputs. This
aligns with what is described in the `action.yml` file. Also adjust the
output names to align with common naming practices. These changes will allow
the program versions to be accessed like so:

steps.<id of step using this Action>.outputs.terraform-version