Convert to `composite` Action
This converts this Action to a `composite` Action from a `docker` Action.
Since we only use this Action to provide version through outputs we do not
need to go through the trouble of building and running a Docker image now
that there are `composite` Actions:
https://docs.github.com/en/actions/creating-actions/creating-a-composite-action

This should simplify this Action greatly and reduce its run time.